requirements.txt
click==8.1.7
pillow==10.1.0
pyserial==3.5
qrcode==8.0.0